# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Opsy Backend Backend (in GoLang)",
        "title": "Opsy Backend",
        "contact": {
            "name": "Priyanshu Rai"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.0.1"
    },
    "host": "nssdmvtpj2.us-east-1.awsapprunner.com",
    "basePath": "/",
    "paths": {
...
        "/user/resend-otp": {
            "post": {
                "description": "Resend 6 digit OTP to email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user authorization"
                ],
                "summary": "Resend OTP",
                "parameters": [
                    {
                        "description": "Resend 6 digit OTP to email",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userAuth.ResendOtpReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userAuth.UserPasswordResDto"
                        }
                    }
                }
            }
        },
...
        "/user/signup": {
            "post": {
                "description": "Signup user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user authorization"
                ],
                "summary": "Signup user",
                "parameters": [
                    {
                        "description": "send 6 digit otp to email for signup",
                        "name": "signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userAuth.SignupReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userAuth.SignupResDto"
                        }
                    }
                }
            }
        },
...
        "/user/verify-otp-for-signup": {
            "post": {
                "description": "Verify the entered 6 digit OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user authorization"
                ],
                "summary": "Verify OTP for signup",
                "parameters": [
                    {
                        "description": "Verify 6 digit OTP and insert data into database",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userAuth.VerifyOtpSignupReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userAuth.VerifyOtpSignupResDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "userAuth.ResendOtpReqDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "userAuth.SignupReqDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "userAuth.SignupResDto": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "userAuth.UserPasswordResDto": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "userAuth.UserResDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "userAuth.VerifyOtpSignupReqDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "otp": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "userAuth.VerifyOtpSignupResDto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/userAuth.UserResDto"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    }
}